+++ config/config.go
@@ -16,10 +16,12 @@ limitations under the License.
 package config
 
 import (
+	"net"
 	"os"
 	"regexp"
 	"strings"
 
+	"github.com/ClickHouse/clickhouse-go/v2"
 	"github.com/hjson/hjson-go/v4"
 	"go.uber.org/zap"
 
@@ -44,9 +46,10 @@ type Config struct {
 
 // KafkaConfig configuration parameters
 type KafkaConfig struct {
-	Brokers  string
-	Security map[string]string
-	TLS      struct {
+	Brokers        string
+	ResetSaslRealm bool
+	Security       map[string]string
+	TLS            struct {
 		Enable         bool
 		CaCertFiles    string // CA cert.pem with which Kafka brokers certs be signed.  Leave empty for certificates trusted by the OS
 		ClientCertFile string // Required for client authentication. It's client cert.pem.
@@ -92,6 +95,7 @@ type ClickHouseConfig struct {
 	Port     int
 	Username string
 	Password string
+	Protocol string //native, http
 
 	// Whether enable TLS encryption with clickhouse-server
 	Secure bool
@@ -139,6 +143,8 @@ type TaskConfig struct {
 		WhiteList string // the regexp of white list
 		BlackList string // the regexp of black list
 	}
+	// additional fields to be appended to each input message, should be a valid json string
+	Fields string `json:"fields,omitempty"`
 	// PrometheusSchema expects each message is a Prometheus metric(timestamp, value, metric name and a list of labels).
 	PrometheusSchema bool
 	// fields match PromLabelsBlackList are not considered as labels. Requires PrometheusSchema be true.
@@ -246,12 +252,29 @@ func (cfg *Config) Normallize(constructGroup bool, httpAddr string, cred util.Cr
 			err = errors.Newf("kafka SASL mechanism %s is unsupported", cfg.Kafka.Sasl.Mechanism)
 			return
 		}
+
+		if cfg.Kafka.ResetSaslRealm {
+			port := getKfkPort(cfg.Kafka.Brokers)
+			os.Setenv("DOMAIN_REALM", net.JoinHostPort("hadoop."+strings.ToLower(cfg.Kafka.Sasl.GSSAPI.Realm), port))
+		}
 	}
 	if cfg.Clickhouse.RetryTimes < 0 {
 		cfg.Clickhouse.RetryTimes = 0
 	}
 	if cfg.Clickhouse.MaxOpenConns <= 0 {
 		cfg.Clickhouse.MaxOpenConns = defaultMaxOpenConns
+	}
+
+	if cfg.Clickhouse.Protocol == "" {
+		cfg.Clickhouse.Protocol = clickhouse.Native.String()
+	}
+
+	if cfg.Clickhouse.Port == 0 {
+		if cfg.Clickhouse.Protocol == clickhouse.HTTP.String() {
+			cfg.Clickhouse.Port = 8123
+		} else {
+			cfg.Clickhouse.Port = 9000
+		}
 	}
 
 	if cfg.Task != nil {
@@ -391,45 +414,44 @@ func (cfg *Config) convertKfkSecurity() {
 	}
 
 	if strings.Contains(protocol, "SSL") {
-		cfg.Kafka.TLS.Enable = true
-		cfg.Kafka.TLS.EndpIdentAlgo = cfg.Kafka.Security["ssl.endpoint.identification.algorithm"]
-		cfg.Kafka.TLS.TrustStoreLocation = cfg.Kafka.Security["ssl.truststore.location"]
-		cfg.Kafka.TLS.TrustStorePassword = cfg.Kafka.Security["ssl.truststore.password"]
-		cfg.Kafka.TLS.KeystoreLocation = cfg.Kafka.Security["ssl.keystore.location"]
-		cfg.Kafka.TLS.KeystorePassword = cfg.Kafka.Security["ssl.keystore.password"]
+		util.TrySetValue(&cfg.Kafka.TLS.Enable, true)
+		util.TrySetValue(&cfg.Kafka.TLS.EndpIdentAlgo, cfg.Kafka.Security["ssl.endpoint.identification.algorithm"])
+		util.TrySetValue(&cfg.Kafka.TLS.TrustStoreLocation, cfg.Kafka.Security["ssl.truststore.location"])
+		util.TrySetValue(&cfg.Kafka.TLS.TrustStorePassword, cfg.Kafka.Security["ssl.truststore.password"])
+		util.TrySetValue(&cfg.Kafka.TLS.KeystoreLocation, cfg.Kafka.Security["ssl.keystore.location"])
+		util.TrySetValue(&cfg.Kafka.TLS.KeystorePassword, cfg.Kafka.Security["ssl.keystore.password"])
 	}
 
 	if strings.Contains(protocol, "SASL") {
-		cfg.Kafka.Sasl.Enable = true
-		cfg.Kafka.Sasl.Mechanism = cfg.Kafka.Security["sasl.mechanism"]
+		util.TrySetValue(&cfg.Kafka.Sasl.Enable, true)
+		util.TrySetValue(&cfg.Kafka.Sasl.Mechanism, cfg.Kafka.Security["sasl.mechanism"])
 		if config, ok := cfg.Kafka.Security["sasl.jaas.config"]; ok {
 			configMap := readConfig(config)
 			if strings.Contains(cfg.Kafka.Sasl.Mechanism, "GSSAPI") {
 				// GSSAPI
 				if configMap["useKeyTab"] != "true" {
-					// Username and password
-					cfg.Kafka.Sasl.GSSAPI.AuthType = 1
-					cfg.Kafka.Sasl.GSSAPI.Username = configMap["username"]
-					cfg.Kafka.Sasl.GSSAPI.Password = configMap["password"]
+					//Username and password
+					util.TrySetValue(&cfg.Kafka.Sasl.GSSAPI.AuthType, 1)
+					util.TrySetValue(&cfg.Kafka.Sasl.GSSAPI.Username, configMap["username"])
+					util.TrySetValue(&cfg.Kafka.Sasl.GSSAPI.Password, configMap["password"])
 				} else {
-					// Keytab
-					cfg.Kafka.Sasl.GSSAPI.AuthType = 2
-					cfg.Kafka.Sasl.GSSAPI.KeyTabPath = configMap["keyTab"]
+					//Keytab
+					util.TrySetValue(&cfg.Kafka.Sasl.GSSAPI.AuthType, 2)
+					util.TrySetValue(&cfg.Kafka.Sasl.GSSAPI.KeyTabPath, configMap["keyTab"])
 					if principal, ok := configMap["principal"]; ok {
-						username := strings.Split(principal, "@")[0]
-						realm := strings.Split(principal, "@")[1]
-						cfg.Kafka.Sasl.GSSAPI.Username = username
-						cfg.Kafka.Sasl.GSSAPI.Realm = realm
+						prins := strings.Split(principal, "@")
+						util.TrySetValue(&cfg.Kafka.Sasl.GSSAPI.Username, prins[0])
+						if len(prins) > 1 {
+							util.TrySetValue(&cfg.Kafka.Sasl.GSSAPI.Realm, prins[1])
+						}
 					}
-					cfg.Kafka.Sasl.GSSAPI.ServiceName = cfg.Kafka.Security["sasl.kerberos.service.name"]
-					if cfg.Kafka.Sasl.GSSAPI.KerberosConfigPath == "" {
-						cfg.Kafka.Sasl.GSSAPI.KerberosConfigPath = defaultKerberosConfigPath
-					}
+					util.TrySetValue(&cfg.Kafka.Sasl.GSSAPI.ServiceName, cfg.Kafka.Security["sasl.kerberos.service.name"])
+					util.TrySetValue(&cfg.Kafka.Sasl.GSSAPI.KerberosConfigPath, defaultKerberosConfigPath)
 				}
 			} else {
 				// PLAIN, SCRAM-SHA-256 or SCRAM-SHA-512
-				cfg.Kafka.Sasl.Username = configMap["username"]
-				cfg.Kafka.Sasl.Password = configMap["password"]
+				util.TrySetValue(&cfg.Kafka.Sasl.Username, configMap["username"])
+				util.TrySetValue(&cfg.Kafka.Sasl.Password, configMap["password"])
 			}
 		}
 	}
@@ -460,4 +482,17 @@ func readConfig(config string) map[string]string {
 		}
 	}
 	return configMap
+}
+
+func getKfkPort(brokers string) string {
+	hosts := strings.Split(brokers, ",")
+	var port string
+	for _, host := range hosts {
+		_, p, err := net.SplitHostPort(host)
+		if err != nil {
+			port = p
+			break
+		}
+	}
+	return port
 }