+++ pool/conn.go
@@ -19,12 +19,14 @@ package pool
 // Clickhouse connection pool
 
 import (
+	"context"
 	"crypto/tls"
 	"fmt"
 	"sync"
 	"time"
 
 	"github.com/ClickHouse/clickhouse-go/v2"
+	"github.com/housepower/clickhouse_sinker/config"
 	"github.com/housepower/clickhouse_sinker/util"
 	"github.com/thanos-io/thanos/pkg/errors"
 	"go.uber.org/zap"
@@ -38,12 +40,14 @@ var (
 // ShardConn a datastructure for storing the clickhouse connection
 type ShardConn struct {
 	lock        sync.Mutex
-	db          clickhouse.Conn
+	conn        *Conn
 	dbVer       int
 	opts        clickhouse.Options
 	replicas    []string         //ip:port list of replicas
 	nextRep     int              //index of next replica
 	writingPool *util.WorkerPool //the all tasks' writing ClickHouse, cpu-net balance
+	protocol    clickhouse.Protocol
+	chCfg       *config.ClickHouseConfig
 }
 
 func (sc *ShardConn) SubmitTask(fn func()) (err error) {
@@ -54,7 +58,7 @@ func (sc *ShardConn) SubmitTask(fn func()) (err error) {
 func (sc *ShardConn) GetReplica() (replica string) {
 	sc.lock.Lock()
 	defer sc.lock.Unlock()
-	if sc.db != nil {
+	if sc.conn != nil {
 		curRep := (len(sc.replicas) + sc.nextRep - 1) % len(sc.replicas)
 		replica = sc.replicas[curRep]
 	}
@@ -65,9 +69,9 @@ func (sc *ShardConn) GetReplica() (replica string) {
 func (sc *ShardConn) Close() {
 	sc.lock.Lock()
 	defer sc.lock.Unlock()
-	if sc.db != nil {
-		sc.db.Close()
-		sc.db = nil
+	if sc.conn != nil {
+		sc.conn.Close()
+		sc.conn = nil
 	}
 	if sc.writingPool != nil {
 		sc.writingPool.StopWait()
@@ -75,10 +79,10 @@ func (sc *ShardConn) Close() {
 }
 
 // NextGoodReplica connects to next good replica
-func (sc *ShardConn) NextGoodReplica(failedVer int) (db clickhouse.Conn, dbVer int, err error) {
+func (sc *ShardConn) NextGoodReplica(failedVer int) (db *Conn, dbVer int, err error) {
 	sc.lock.Lock()
 	defer sc.lock.Unlock()
-	if sc.db != nil {
+	if sc.conn != nil {
 		if sc.dbVer > failedVer {
 			// Another goroutine has already done connection.
 			// Notice: Why recording failure version instead timestamp?
@@ -86,70 +90,91 @@ func (sc *ShardConn) NextGoodReplica(failedVer int) (db clickhouse.Conn, dbVer i
 			// conn1 = NextGood(0); conn2 = NexGood(0); conn1.Exec failed at ts1;
 			// conn3 = NextGood(ts1); conn2.Exec failed at ts2;
 			// conn4 = NextGood(ts2) will close the good connection and break users.
-			return sc.db, sc.dbVer, nil
+			return sc.conn, sc.dbVer, nil
 		}
-		sc.db.Close()
-		sc.db = nil
+		sc.conn.Close()
+		sc.conn = nil
 	}
 	savedNextRep := sc.nextRep
 	// try all replicas, including the current one
+	conn := Conn{
+		protocol: sc.protocol,
+		ctx:      context.Background(),
+	}
 	for i := 0; i < len(sc.replicas); i++ {
 		replica := sc.replicas[sc.nextRep]
 		sc.opts.Addr = []string{replica}
 		sc.nextRep = (sc.nextRep + 1) % len(sc.replicas)
-		sc.db, err = clickhouse.Open(&sc.opts)
+		if sc.protocol == clickhouse.HTTP {
+			conn.db = clickhouse.OpenDB(&sc.opts)
+			conn.db.SetMaxOpenConns(sc.chCfg.MaxOpenConns)
+			conn.db.SetMaxIdleConns(sc.chCfg.MaxOpenConns)
+			conn.db.SetConnMaxLifetime(time.Minute * 10)
+		} else {
+			conn.c, err = clickhouse.Open(&sc.opts)
+		}
 		if err != nil {
 			util.Logger.Warn("clickhouse.Open failed", zap.String("replica", replica), zap.Error(err))
 			continue
 		}
 		sc.dbVer++
 		util.Logger.Info("clickhouse.Open succeeded", zap.Int("dbVer", sc.dbVer), zap.String("replica", replica))
-		return sc.db, sc.dbVer, nil
+		sc.conn = &conn
+		return sc.conn, sc.dbVer, nil
 	}
 	err = errors.Newf("no good replica among replicas %v since %d", sc.replicas, savedNextRep)
 	return nil, sc.dbVer, err
 }
 
-// Each shard has a clickhouse.Conn which connects to one replica inside the shard.
+// Each shard has a pool.Conn which connects to one replica inside the shard.
 // We need more control than replica single-point-failure.
-func InitClusterConn(hosts [][]string, port int, db, username, password string, secure, skipVerify bool,
-	maxOpenConns int) (err error) {
+func InitClusterConn(chCfg *config.ClickHouseConfig) (err error) {
 	lock.Lock()
 	defer lock.Unlock()
 	freeClusterConn()
 
-	for _, replicas := range hosts {
+	proto := clickhouse.Native
+	if chCfg.Protocol == clickhouse.HTTP.String() {
+		proto = clickhouse.HTTP
+	}
+
+	for _, replicas := range chCfg.Hosts {
 		numReplicas := len(replicas)
 		replicaAddrs := make([]string, numReplicas)
 		for i, ip := range replicas {
 			// Changing hostnames to IPs breaks TLS connections in many cases
-			if !secure {
+			if !chCfg.Secure {
 				if ips2, err := util.GetIP4Byname(ip); err == nil {
 					ip = ips2[0]
 				}
 			}
-			replicaAddrs[i] = fmt.Sprintf("%s:%d", ip, port)
+			replicaAddrs[i] = fmt.Sprintf("%s:%d", ip, chCfg.Port)
 		}
 		sc := &ShardConn{
 			replicas: replicaAddrs,
+			chCfg:    chCfg,
 			opts: clickhouse.Options{
 				Auth: clickhouse.Auth{
-					Database: db,
-					Username: username,
-					Password: password,
+					Database: chCfg.DB,
+					Username: chCfg.Username,
+					Password: chCfg.Password,
 				},
-				DialTimeout:     time.Minute * 10,
-				MaxOpenConns:    maxOpenConns,
-				MaxIdleConns:    5, // TODO - update this property to maxOpenConns when the lifetime of an idle connection honours the ConnMaxLifetime
-				ConnMaxLifetime: time.Minute * 10,
+				Protocol:    proto,
+				DialTimeout: time.Minute * 10,
 			},
-			writingPool: util.NewWorkerPool(maxOpenConns, 1),
+			writingPool: util.NewWorkerPool(chCfg.MaxOpenConns, 1),
 		}
-		if secure {
+		if chCfg.Secure {
 			tlsConfig := &tls.Config{}
-			tlsConfig.InsecureSkipVerify = skipVerify
+			tlsConfig.InsecureSkipVerify = chCfg.InsecureSkipVerify
 			sc.opts.TLS = tlsConfig
 		}
+		if proto == clickhouse.Native {
+			sc.opts.MaxOpenConns = chCfg.MaxOpenConns
+			sc.opts.MaxIdleConns = chCfg.MaxOpenConns
+			sc.opts.ConnMaxLifetime = time.Minute * 10
+		}
+		sc.protocol = proto
 		if _, _, err = sc.NextGoodReplica(0); err != nil {
 			return
 		}