+++ util/common.go
@@ -20,8 +20,6 @@ import (
 	"crypto/tls"
 	"crypto/x509"
 	"fmt"
-	"math"
-	"net"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -98,51 +96,6 @@ func GetSourceName(parser, name string) (sourcename string) {
 // GetShift returns the smallest `shift` which 1<<shift is no smaller than s
 func GetShift(s int) (shift uint) {
 	for shift = 0; (1 << shift) < s; shift++ {
-	}
-	return
-}
-
-// GetOutboundIP gets preferred outbound ip of this machine
-// https://stackoverflow.com/questions/23558425/how-do-i-get-the-local-ip-address-in-go.
-func GetOutboundIP() (ip net.IP, err error) {
-	var conn net.Conn
-	if conn, err = net.Dial("udp", "8.8.8.8:80"); err != nil {
-		err = errors.Wrapf(err, "")
-		return
-	}
-	defer conn.Close()
-	localAddr, _ := conn.LocalAddr().(*net.UDPAddr)
-	ip = localAddr.IP
-	return
-}
-
-// GetSpareTCPPort finds a spare TCP port.
-func GetSpareTCPPort(portBegin int) int {
-	for port := portBegin; port < math.MaxInt; port++ {
-		if err := testListenOnPort(port); err == nil {
-			return port
-		}
-	}
-	return 0
-}
-
-func testListenOnPort(port int) error {
-	addr := fmt.Sprintf(":%d", port)
-	ln, err := net.Listen("tcp", addr)
-	if err != nil {
-		return err
-	}
-	ln.Close() //nolint:errcheck
-	return nil
-}
-
-// https://stackoverflow.com/questions/50428176/how-to-get-ip-and-port-from-net-addr-when-it-could-be-a-net-udpaddr-or-net-tcpad
-func GetNetAddrPort(addr net.Addr) (port int) {
-	switch addr := addr.(type) {
-	case *net.UDPAddr:
-		port = addr.Port
-	case *net.TCPAddr:
-		port = addr.Port
 	}
 	return
 }
@@ -298,4 +251,69 @@ func SetLogLevel(newLogLevel string) {
 		}
 		logAtomLevel.SetLevel(lvl)
 	}
+}
+
+// set v2 to v1, if v1 didn't bind any value
+// FIXME: how about v1 bind default value?
+func TrySetValue(v1, v2 interface{}) bool {
+	var ok bool
+	rt := reflect.TypeOf(v1)
+	rv := reflect.ValueOf(v1)
+
+	if rt.Kind() != reflect.Ptr {
+		return ok
+	}
+	for rt.Kind() == reflect.Ptr {
+		rt = rt.Elem()
+		rv = rv.Elem()
+	}
+
+	if rv.IsValid() && rv.IsZero() {
+		ok = true
+		switch rt.Kind() {
+		case reflect.Uint:
+			v, _ := v2.(uint)
+			rv.SetUint(uint64(v))
+		case reflect.Uint8:
+			v, _ := v2.(uint8)
+			rv.SetUint(uint64(v))
+		case reflect.Uint16:
+			v, _ := v2.(uint16)
+			rv.SetUint(uint64(v))
+		case reflect.Uint32:
+			v, _ := v2.(uint32)
+			rv.SetUint(uint64(v))
+		case reflect.Uint64:
+			v, _ := v2.(uint64)
+			rv.SetUint(uint64(v))
+		case reflect.Int:
+			v, _ := v2.(int)
+			rv.SetInt(int64(v))
+		case reflect.Int8:
+			v, _ := v2.(int8)
+			rv.SetInt(int64(v))
+		case reflect.Int16:
+			v, _ := v2.(int16)
+			rv.SetInt(int64(v))
+		case reflect.Int32:
+			v, _ := v2.(int32)
+			rv.SetInt(int64(v))
+		case reflect.Int64:
+			v, _ := v2.(int64)
+			rv.SetInt(int64(v))
+		case reflect.Float32:
+			v, _ := v2.(float32)
+			rv.SetFloat(float64(v))
+		case reflect.Float64:
+			v, _ := v2.(float64)
+			rv.SetFloat(float64(v))
+		case reflect.String:
+			rv.SetString(v2.(string))
+		case reflect.Bool:
+			rv.SetBool(v2.(bool))
+		default:
+			ok = false
+		}
+	}
+	return ok
 }